% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cov_pca.R
\name{cov_pca}
\alias{cov_pca}
\title{Find the Top k Eigenvalues and Eigenvectors of the Covariance Matrix}
\usage{
cov_pca(x, k, w = 1, covmat = cov(sweep(x, 1, rowMeans(x, na.rm = TRUE),
  "-"), use = "pairwise"))
}
\arguments{
\item{x}{A numeric matrix, possibly with missing values.}

\item{k}{Number of eigenvalue/eigenvector pairs to return.}

\item{w}{A vector of weights for the columns of \code{x}.}

\item{covmat}{The covariance matrix of \code{x}, if pre-computed (say, with \code{running_cov}).}
}
\value{
A numeric vector: 
the first entry gives the total sum of squared values of the covariance matrix 
(i.e., the sum of the eigenvectors, for computing the proportion of variance explained);
the next k entires give eigenvalues and each subsequent group of ncol(data) columns give the corresponding eigenvector.

If C is the covariance matrix, the columns of U are the eigenvectors, and W and Lambda are diagonal matrices with w and the eigenvalues on the diagonals respectively,
then the output is the minimizer of
      | W^(1/2) ( C - U Lambda U' ) W^(1/2) |^2 = sum_{ij} w[i] w[j] ( C[i,j] - (U Lambda U')[i,j] )^2 ,
and the sum-of-squares (equal to the sum of the eigenvalues squared) is
      sum_{ij}  w[i] * w[j] * C[i,j]^2 .
}
\description{
This takes a numeric matrix, subtracts row means, finds the covariance matrix of the columns (using \code{use="pairwise"}),
and returns a vector containing: the sum of squared entries in the covariance matrix; the top k eigenvalues, and the top k eigenvectors.
}

