% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_tped}
\alias{read_tped}
\title{Read Chromosome(s) From a TPED File}
\usage{
read_tped(file, chrom = "[^ \\t]*", triallelic = TRUE, phased = FALSE)
}
\arguments{
\item{file}{Input .tped filename.}

\item{chrom}{Vector of chromosome name(s) to include (defaults to all).}

\item{triallelic}{Include triallelic sites? Defaults to TRUE; otherwise sets these to NA.}

\item{phased}{Are the data phased? (i.e., should return one column per individual or two?)}
}
\value{
An integer matrix with one row per site and one (or two, if phased) column per individual.
}
\description{
Reads specified chromsomes from a .tped file and recodes the data into numeric format, e.g. for diploids:
   0 = major homozygote
   1 = heterozygote
   2 = minor homozygote
where "major" is the most common allele at that site, and all others are lumped as "minor".
Assumes that anything not in "ACGT" is NA (case-sensitive).
}
\details{
A .tped has one row per SNP, *two* columns per sample.  The first column is chromosome, the fourth is position.
Does no error checking: let the user beware.
}

