% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_data.R
\name{read_vcf}
\alias{read_vcf}
\title{Read Chromosome(s) From a VCF File}
\usage{
read_vcf(file, phased = FALSE, triallelic = TRUE)
}
\arguments{
\item{file}{Input .vcf file, possibly gzip'ped.}

\item{phased}{Are the data phased? (i.e., should return one column per individual or two?)}

\item{triallelic}{Include triallelic sites? Defaults to TRUE.}
}
\value{
An integer matrix with one row per site and one column per individual.
}
\description{
Reads specified chromsomes from a .vcf file and recodes the data into numeric format:
   0 = major homozygote
   1 = heterozygote
   2 = minor homozygote
where "major" is the most common allele at that site, and all others are lumped as "minor".
Assumes that anything not in "ACGT" is NA (case-sensitive).
}

