% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf.R
\name{region_string}
\alias{region_string}
\title{Make Regions String from Data Frame of Regions}
\usage{
region_string(regions)
}
\arguments{
\item{regions}{A data frame with \code{chrom}, \code{start}, and \code{end}.}
}
\value{
A character string of the form "chr1:start1-end1,chr2:start2,end2".
}
\description{
Given a data frame with variables \code{chrom}, \code{start}, and \code{end},
make a regions string that can be passed to bcftools.
}

