% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/query_vcf.R
\name{vcf_samples}
\alias{vcf_samples}
\title{Find Sample IDs from a VCF File}
\usage{
vcf_samples(file)
}
\arguments{
\item{file}{The name of the indexed vcf file.}
}
\value{
A character vector of sample IDs.
}
\description{
Returns the character vector of samples (by calling \code{bcftools query -l}).
}

