% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corners.R
\name{corners}
\alias{corners}
\title{Find Points in a 2D Cloud Closest to the "Corners"}
\usage{
corners(xy, prop, k = 3)
}
\arguments{
\item{xy}{A two-column numeric matrix of coordinates.}

\item{prop}{The proportion of points to return for each corner.}
}
\value{
A three column integer matrix giving the indices of the corresponding points.
}
\description{
Finds the k "extreme points" that lie closest to the minimal enclosing circle,
and returns a matrix with k columns containing the portion of points 
that lie closest to those extreme points.
If k is larger than the number of points on the minimal enclosing circle,
adds the closest points to that circle that are not in previously added points' neighborhoods
until k is reached.
}

